+++ EatFitService/NutritionService/admin.py
@@ -61,6 +61,11 @@ class NutrientInline(admin.TabularInline):
 
 class NutriScoreFactsInline(admin.StackedInline):
     model = NutriScoreFacts
+    readonly_fields = ('total_fvpn_percentage', 'ofcom_n_energy_kj', 'ofcom_n_saturated_fat', 'ofcom_n_sugars',
+                       'ofcom_n_salt', 'ofcom_p_protein', 'ofcom_p_fvpn', 'ofcom_p_dietary_fiber',
+                       'ofcom_n_energy_kj_mixed', 'ofcom_n_saturated_fat_mixed', 'ofcom_n_sugars_mixed',
+                       'ofcom_n_salt_mixed', 'ofcom_p_protein_mixed', 'ofcom_p_fvpn_mixed',
+                       'ofcom_p_dietary_fiber_mixed')
 
 
 class AdditionalImageInline(admin.StackedInline):
@@ -87,6 +92,8 @@ class MarketRegionInline(admin.TabularInline):
 class ProductAdmin(admin.ModelAdmin):
     list_display = ("product_name_de", "gtin")
     search_fields = ('product_name_de', 'gtin')
+    readonly_fields = ('major_category', 'nutri_score_final', 'nutri_score_calculated', 'nutri_score_calculated_mixed',
+                       'nutri_score_number_of_errors', 'ofcom_value', 'found_count')
     save_as = True
     inlines = [
         NutriScoreFactsInline,